+++ server/src/s_challenges.ts
@@ -11,8 +11,17 @@ export class CSubmission{
     who:string;
     uid:string;
     pid:string;
+    cc = 0; // char count
+    lang:string[] = [];
+    ws = ""; // when submitted
+    t = 0; // time taken
     static from(data:any){
-        return new CSubmission(data.url,data.who,data.uid,data.pid);
+        let c = new CSubmission(data.url,data.who,data.uid,data.pid);
+        c.cc = data.cc ?? 0;
+        c.lang = data.lang ?? [];
+        c.ws = data.ws ?? new Date().toISOString();
+        c.t = data.t ?? 0;
+        return c;
     }
 }
 export class Challenge{